pyproject.toml
[project]
name = "gcp-mcp-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-auth>=2.48.0",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
]
